[MODDICONV-295] Disallow deletion of default MARC field protection settings

+++ mod-di-converter-storage-server/src/main/java/org/folio/services/fieldprotectionsettings/MarcFieldProtectionSettingsServiceImpl.java
@@ -36,17 +36,29 @@ public class MarcFieldProtectionSettingsServiceImpl implements MarcFieldProtecti
 
   @Override
   public Future<MarcFieldProtectionSetting> updateMarcFieldProtectionSetting(MarcFieldProtectionSetting marcFieldProtectionSetting, String tenantId) {
-    if (MarcFieldProtectionSetting.Source.SYSTEM == marcFieldProtectionSetting.getSource()) {
-      return Future.failedFuture(new BadRequestException("MARC field protection setting with source SYSTEM cannot be updated"));
-    }
     return getMarcFieldProtectionSettingById(marcFieldProtectionSetting.getId(), tenantId)
-      .compose(optionalSetting -> optionalSetting
-        .map(setting -> fieldProtectionSettingsDao.update(marcFieldProtectionSetting, tenantId))
-        .orElse(Future.failedFuture(new NotFoundException(String.format("MARC field protection with id '%s' was not found", marcFieldProtectionSetting.getId())))));
+      .compose(optionalSetting -> {
+        if (optionalSetting.isPresent()) {
+          if (MarcFieldProtectionSetting.Source.SYSTEM == optionalSetting.get().getSource()) {
+            return Future.failedFuture(new BadRequestException("MARC field protection setting with source SYSTEM cannot be updated"));
+          }
+          return fieldProtectionSettingsDao.update(marcFieldProtectionSetting, tenantId);
+        }
+        return Future.failedFuture(new NotFoundException(String.format("MARC field protection with id '%s' was not found", marcFieldProtectionSetting.getId())));
+      });
   }
 
   @Override
   public Future<Boolean> deleteMarcFieldProtectionSetting(String id, String tenantId) {
-    return fieldProtectionSettingsDao.delete(id, tenantId);
+    return getMarcFieldProtectionSettingById(id, tenantId)
+      .compose(optionalSetting -> {
+        if (optionalSetting.isPresent()) {
+          if (MarcFieldProtectionSetting.Source.SYSTEM == optionalSetting.get().getSource()) {
+            return Future.failedFuture(new BadRequestException("MARC field protection setting with source SYSTEM cannot be deleted"));
+          }
+          return fieldProtectionSettingsDao.delete(id, tenantId);
+        }
+        return Future.succeededFuture(Boolean.FALSE);
+      });
   }
 }
